refactor navier_stokes_calculators; compressible Q-Criterion
