+++ include/exadg/compressible_navier_stokes/spatial_discretization/calculators.h
@@ -33,149 +33,6 @@ namespace ExaDG
 {
 namespace CompNS
 {
-template<int dim, typename Number>
-class VorticityCalculator
-{
-public:
-  static unsigned int const number_vorticity_components = (dim == 2) ? 1 : dim;
-
-  typedef dealii::LinearAlgebra::distributed::Vector<Number> VectorType;
-
-  typedef VorticityCalculator<dim, Number> This;
-
-  typedef CellIntegrator<dim, dim, Number> CellIntegratorVector;
-
-  typedef dealii::Tensor<1, dim, dealii::VectorizedArray<Number>> vector;
-
-  VorticityCalculator() : matrix_free(nullptr), dof_index(0), quad_index(0){};
-
-  void
-  initialize(dealii::MatrixFree<dim, Number> const & matrix_free_in,
-             unsigned int const                      dof_index_in,
-             unsigned int const                      quad_index_in)
-  {
-    matrix_free = &matrix_free_in;
-    dof_index   = dof_index_in;
-    quad_index  = quad_index_in;
-  }
-
-  void
-  compute_vorticity(VectorType & dst, VectorType const & src) const
-  {
-    matrix_free->cell_loop(&This::local_compute_vorticity, this, dst, src);
-  }
-
-private:
-  void
-  local_compute_vorticity(dealii::MatrixFree<dim, Number> const &       matrix_free,
-                          VectorType &                                  dst,
-                          VectorType const &                            src,
-                          std::pair<unsigned int, unsigned int> const & cell_range) const
-  {
-    CellIntegratorVector integrator(matrix_free, dof_index, quad_index, 0);
-
-    for(unsigned int cell = cell_range.first; cell < cell_range.second; ++cell)
-    {
-      integrator.reinit(cell);
-      integrator.read_dof_values(src);
-      integrator.evaluate(dealii::EvaluationFlags::gradients);
-
-      for(unsigned int q = 0; q < integrator.n_q_points; ++q)
-      {
-        dealii::Tensor<1, number_vorticity_components, dealii::VectorizedArray<Number>> omega =
-          integrator.get_curl(q);
-
-        // omega_vector is a vector with dim components
-        // for dim=3: omega_vector[i] = omega[i], i=1,...,dim
-        // for dim=2: omega_vector[0] = omega,
-        //            omega_vector[1] = 0
-        vector omega_vector;
-        for(unsigned int d = 0; d < number_vorticity_components; ++d)
-          omega_vector[d] = omega[d];
-        integrator.submit_value(omega_vector, q);
-      }
-
-      integrator.integrate(dealii::EvaluationFlags::values);
-      integrator.set_dof_values(dst);
-    }
-  }
-
-  dealii::MatrixFree<dim, Number> const * matrix_free;
-  unsigned int                            dof_index;
-  unsigned int                            quad_index;
-};
-
-
-template<int dim, typename Number>
-class DivergenceCalculator
-{
-public:
-  typedef dealii::LinearAlgebra::distributed::Vector<Number> VectorType;
-
-  typedef DivergenceCalculator<dim, Number> This;
-
-  typedef CellIntegrator<dim, dim, Number> CellIntegratorVector;
-  typedef CellIntegrator<dim, 1, Number>   CellIntegratorScalar;
-
-  typedef dealii::VectorizedArray<Number> scalar;
-
-  DivergenceCalculator()
-    : matrix_free(nullptr), dof_index_vector(1), dof_index_scalar(2), quad_index(0){};
-
-  void
-  initialize(dealii::MatrixFree<dim, Number> const & matrix_free_in,
-             unsigned int const                      dof_index_vector_in,
-             unsigned int const                      dof_index_scalar_in,
-             unsigned int const                      quad_index_in)
-  {
-    matrix_free      = &matrix_free_in;
-    dof_index_vector = dof_index_vector_in;
-    dof_index_scalar = dof_index_scalar_in;
-    quad_index       = quad_index_in;
-  }
-
-  void
-  compute_divergence(VectorType & dst, VectorType const & src) const
-  {
-    matrix_free->cell_loop(&This::local_compute_divergence, this, dst, src);
-  }
-
-private:
-  void
-  local_compute_divergence(dealii::MatrixFree<dim, Number> const &       matrix_free,
-                           VectorType &                                  dst,
-                           VectorType const &                            src,
-                           std::pair<unsigned int, unsigned int> const & cell_range) const
-  {
-    CellIntegratorVector integrator_vector(matrix_free, dof_index_vector, quad_index, 0);
-    CellIntegratorScalar integrator_scalar(matrix_free, dof_index_scalar, quad_index, 0);
-
-    for(unsigned int cell = cell_range.first; cell < cell_range.second; ++cell)
-    {
-      integrator_vector.reinit(cell);
-      integrator_vector.read_dof_values(src);
-      integrator_vector.evaluate(dealii::EvaluationFlags::gradients);
-
-      integrator_scalar.reinit(cell);
-
-      for(unsigned int q = 0; q < integrator_vector.n_q_points; q++)
-      {
-        scalar divergence = integrator_vector.get_divergence(q);
-        integrator_scalar.submit_value(divergence, q);
-      }
-
-      integrator_scalar.integrate(dealii::EvaluationFlags::values);
-      integrator_scalar.set_dof_values(dst);
-    }
-  }
-
-  dealii::MatrixFree<dim, Number> const * matrix_free;
-
-  unsigned int dof_index_vector;
-  unsigned int dof_index_scalar;
-  unsigned int quad_index;
-};
-
 /*
  *  The DoF vector contains the vector of conserved quantities (rho, rho u, rho E).
  *  This class allows to transfer these quantities into the derived variables (p, u, T)
