--- include/exadg/incompressible_navier_stokes/spatial_discretization/calculators/divergence_calculator.h
@@ -1,80 +1,0 @@
-/*  ______________________________________________________________________
- *
- *  ExaDG - High-Order Discontinuous Galerkin for the Exa-Scale
- *
- *  Copyright (C) 2021 by the ExaDG authors
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *  ______________________________________________________________________
- */
-
-#ifndef INCLUDE_EXADG_INCOMPRESSIBLE_NAVIER_STOKES_SPATIAL_DISCRETIZATION_CALCULATORS_DIVERGENCE_CALCULATOR_H_
-#define INCLUDE_EXADG_INCOMPRESSIBLE_NAVIER_STOKES_SPATIAL_DISCRETIZATION_CALCULATORS_DIVERGENCE_CALCULATOR_H_
-
-// deal.II
-#include <deal.II/lac/la_parallel_vector.h>
-
-// ExaDG
-#include <exadg/matrix_free/integrators.h>
-
-namespace ExaDG
-{
-namespace IncNS
-{
-template<int dim, typename Number>
-class DivergenceCalculator
-{
-private:
-  typedef DivergenceCalculator<dim, Number> This;
-
-  typedef dealii::LinearAlgebra::distributed::Vector<Number> VectorType;
-
-  typedef dealii::VectorizedArray<Number> scalar;
-
-  typedef std::pair<unsigned int, unsigned int> Range;
-
-  typedef CellIntegrator<dim, dim, Number> CellIntegratorVector;
-  typedef CellIntegrator<dim, 1, Number>   CellIntegratorScalar;
-
-public:
-  DivergenceCalculator();
-
-  void
-  initialize(dealii::MatrixFree<dim, Number> const & matrix_free_in,
-             unsigned int const                      dof_index_u_in,
-             unsigned int const                      dof_index_u_scalar_in,
-             unsigned int const                      quad_index_in);
-
-  void
-  compute_divergence(VectorType & dst, VectorType const & src) const;
-
-private:
-  void
-  cell_loop(dealii::MatrixFree<dim, Number> const & matrix_free,
-            VectorType &                            dst,
-            VectorType const &                      src,
-            Range const &                           cell_range) const;
-
-  dealii::MatrixFree<dim, Number> const * matrix_free;
-
-  unsigned int dof_index_u;
-  unsigned int dof_index_u_scalar;
-  unsigned int quad_index;
-};
-
-} // namespace IncNS
-} // namespace ExaDG
-
-#endif /* INCLUDE_EXADG_INCOMPRESSIBLE_NAVIER_STOKES_SPATIAL_DISCRETIZATION_CALCULATORS_DIVERGENCE_CALCULATOR_H_ \
-        */
