+++ include/exadg/incompressible_navier_stokes/spatial_discretization/spatial_operator_base.h
@@ -33,11 +33,7 @@
 // ExaDG
 #include <exadg/grid/grid.h>
 #include <exadg/grid/grid_motion_interface.h>
-#include <exadg/incompressible_navier_stokes/spatial_discretization/calculators/divergence_calculator.h>
-#include <exadg/incompressible_navier_stokes/spatial_discretization/calculators/q_criterion_calculator.h>
 #include <exadg/incompressible_navier_stokes/spatial_discretization/calculators/streamfunction_calculator_rhs_operator.h>
-#include <exadg/incompressible_navier_stokes/spatial_discretization/calculators/velocity_magnitude_calculator.h>
-#include <exadg/incompressible_navier_stokes/spatial_discretization/calculators/vorticity_calculator.h>
 #include <exadg/incompressible_navier_stokes/spatial_discretization/operators/convective_operator.h>
 #include <exadg/incompressible_navier_stokes/spatial_discretization/operators/divergence_operator.h>
 #include <exadg/incompressible_navier_stokes/spatial_discretization/operators/gradient_operator.h>
@@ -589,11 +585,11 @@ protected:
   /*
    * Calculators used to obtain derived quantities.
    */
-  VorticityCalculator<dim, Number>         vorticity_calculator;
-  DivergenceCalculator<dim, Number>        divergence_calculator;
-  ShearRateCalculator<dim, Number>         shear_rate_calculator;
-  VelocityMagnitudeCalculator<dim, Number> velocity_magnitude_calculator;
-  QCriterionCalculator<dim, Number>        q_criterion_calculator;
+  VorticityCalculator<dim, Number>  vorticity_calculator;
+  DivergenceCalculator<dim, Number> divergence_calculator;
+  ShearRateCalculator<dim, Number>  shear_rate_calculator;
+  MagnitudeCalculator<dim, Number>  magnitude_calculator;
+  QCriterionCalculator<dim, Number> q_criterion_calculator;
 
   MPI_Comm const mpi_comm;
